src/requirements.txt
MinimalModbus==0.7
paho-mqtt==1.1
pyserial==3.0.1
wheel==0.24.0
